Minimal linter prototype
This implements a prototype Thrift linter based on the AST walking work
done in #285.

Linter rules walk the AST and post messages to the Lint Catcher.
SimpleRule exposes a simplified interface in front of that for basic
linters that don't need control over the AST traversal. These can simply
return true/false on a specific kind of Node and provide a
`text/template` compatible template for the message.